requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Requests==2.31.0
